chore: User Entity 수정
패키징 변경, Unique 속성 설정 및 Setter 삭제
