--- src/main/java/io/zerogone/model/User.java
@@ -1,50 +1,0 @@
-package io.zerogone.model;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-
-@Entity
-public class User {
-    @Id
-    private int id;
-
-    private String name;
-
-    @Column(nullable = false)
-    private String email;
-    @Column(name = "nick_name", nullable = false)
-    private String nickName;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getNickName() {
-        return nickName;
-    }
-
-    public void setNickName(String nickName) {
-        this.nickName = nickName;
-    }
-}
