+++ src/main/java/io/zerogone/user/model/User.java
@@ -1,0 +1,33 @@
+package io.zerogone.user.model;
+
+import javax.persistence.*;
+
+@Entity
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    private String name;
+
+    @Column(nullable = false, unique = true)
+    private String email;
+    @Column(name = "nick_name", nullable = false, unique = true)
+    private String nickName;
+
+    public int getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getNickName() {
+        return nickName;
+    }
+}
